Robocopy: Modest refactoring
