+++ cmd/robocopy/utils.go
@@ -1,6 +1,8 @@
 package main
 
 import (
+	"encoding/json"
+	"fmt"
 	"io"
 	"net/http"
 	"os"
@@ -29,4 +31,18 @@ func readFrom(name string) (rc io.ReadCloser, err error) {
 
 	f, err := os.Open(name)
 	return f, err
+}
+
+func unmarshalFrom(name string, val interface{}) error {
+	rc, err := readFrom(name)
+	if err != nil {
+		return fmt.Errorf("could not read file %q: %v", name, err)
+	}
+	defer deferClose(&err, rc.Close)
+
+	dec := json.NewDecoder(BOMReader(rc))
+	if err = dec.Decode(val); err != nil {
+		return fmt.Errorf("could not decode %q: %v", name, err)
+	}
+	return err
 }
